refactor(@cubejs-client/core): export types and convert type aliases to interfaces
We should be exporting types in general, there's usually no good reason to have them in an `index.d.ts` and then close them off.

Separately, type aliases don't work well with declaration merging, which presents a problem for us as we plan to make use of it. Not all type aliases have been converted; only the ones required for type generation.